+++ src/main/java/com/ybardockz/streaming_music/services/UserService.java
@@ -2,11 +2,14 @@ package com.ybardockz.streaming_music.services;
 
 import java.util.List;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.ybardockz.streaming_music.entities.User;
 import com.ybardockz.streaming_music.repositories.UserRepository;
+import com.ybardockz.streaming_music.services.exceptions.DbException;
+import com.ybardockz.streaming_music.services.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
 
@@ -20,7 +23,7 @@ public class UserService {
 	}
 	
 	public User findById(Long id) {
-		return repository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
+		return repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public List<User> findAll() {
@@ -33,7 +36,15 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DbException(e.getMessage());
+		}
+		
 	}
 	
 	public void update(User user, Long id) {
@@ -45,8 +56,7 @@ public class UserService {
 			
 			repository.save(userExisting);
 		} catch (EntityNotFoundException e) {
-			System.out.println(e.getMessage());
-			e.printStackTrace();
+			throw new ResourceNotFoundException(id);
 		}
